wip initializer poc
