+++ .eslintignore
@@ -4,7 +4,6 @@ scripts/htm.js
 scripts/acdl
 tools/picker
 tools/pdp-metadata
-scripts/dropins
 scripts/__dropins__
 scripts/commerce-events-collector.js
 scripts/commerce-events-sdk.js
