+++ blocks/commerce-create-account/commerce-create-account.js
@@ -5,6 +5,9 @@ import { render as authRenderer } from '@dropins/storefront-auth/render.js';
 import { checkIsAuthenticated } from '../../scripts/configs.js';
 import { CUSTOMER_ACCOUNT_PATH, CUSTOMER_LOGIN_PATH } from '../../scripts/constants.js';
 
+// Initialize
+import '../../scripts/initializers/auth.js';
+
 export default async function decorate(block) {
   if (checkIsAuthenticated()) {
     window.location.href = CUSTOMER_ACCOUNT_PATH;
