+++ blocks/commerce-order-confirmation/commerce-order-confirmation.js
@@ -3,10 +3,8 @@
 
 // Dropin Tools
 import { events } from '@dropins/tools/event-bus.js';
-import { initializers } from '@dropins/tools/initializer.js';
 
 // Order Confirmation Dropin Modules
-import * as orderConfirmationApi from '@dropins/storefront-order-confirmation/api.js';
 import { render as provider } from '@dropins/storefront-order-confirmation/render.js';
 import OrderConfirmation from '@dropins/storefront-order-confirmation/containers/OrderConfirmation.js';
 
@@ -17,12 +15,14 @@ import SignUp from '@dropins/storefront-auth/containers/SignUp.js';
 import { createModal } from '../modal/modal.js';
 import { CUSTOMER_ACCOUNT_PATH, CUSTOMER_LOGIN_PATH, SUPPORT_PATH } from '../../scripts/constants.js';
 
+// Initialize
+import '../../scripts/initializers/auth.js';
+import '../../scripts/initializers/order-confirmation.js';
+
 export default async function decorate(block) {
   let modal = null;
 
   // Initialize Dropins
-  initializers.register(orderConfirmationApi.initialize, {});
-
   events.on('authenticated', (isAuthenticated) => {
     if (isAuthenticated && modal) {
       modal.removeModal();
