+++ blocks/commerce-order-status/commerce-order-status.js
@@ -3,6 +3,9 @@
 import { render as orderRenderer } from '@dropins/storefront-order/render.js';
 import { OrderStatus } from '@dropins/storefront-order/containers/OrderStatus.js';
 
+// Initialize
+import '../../scripts/initializers/order.js';
+
 export default async function decorate(block) {
   await orderRenderer.render(OrderStatus, {})(block);
 }
