+++ blocks/commerce-search-order/commerce-search-order.js
@@ -8,6 +8,10 @@ import { events } from '@dropins/tools/event-bus.js';
 import { checkIsAuthenticated } from '../../scripts/configs.js';
 import { CUSTOMER_ORDER_DETAILS_PATH, ORDER_DETAILS_PATH } from '../../scripts/constants.js';
 
+// Initialize
+import '../../scripts/initializers/auth.js';
+import '../../scripts/initializers/order.js';
+
 const renderSignIn = async (element, email, orderNumber) => authRenderer.render(SignIn, {
   initialEmailValue: email,
   renderSignUpLink: false,
