+++ blocks/commerce-shipping-status/commerce-shipping-status.js
@@ -3,6 +3,9 @@
 import { render as orderRenderer } from '@dropins/storefront-order/render.js';
 import { ShippingStatus } from '@dropins/storefront-order/containers/ShippingStatus.js';
 
+// Initialize
+import '../../scripts/initializers/order.js';
+
 export default async function decorate(block) {
   await orderRenderer.render(ShippingStatus, {})(block);
 }
