+++ head.html
@@ -29,7 +29,7 @@
 <script src="/scripts/configs.js" type="module"></script>
 <script src="/scripts/commerce.js" type="module"></script>
 
-<link rel="modulepreload" href="/scripts/dropins.js" />
+<link rel="modulepreload" href="/scripts/initializers/index.js" />
 <link rel="modulepreload" href="/scripts/__dropins__/tools/event-bus.js" />
 <link rel="modulepreload" href="/scripts/__dropins__/tools/fetch-graphql.js" />
 <link rel="modulepreload" href="/scripts/__dropins__/tools/initializer.js" />
